Calculator
This project is a web-based calculator designed to provide users with a simple and functional tool for performing basic arithmetic calculations. The calculator website is built using HTML, CSS, and JavaScript, offering a clean and responsive interface that ensures a smooth user experience across different devices. The HTML structure defines the layout, the CSS styles enhance the visual presentation, and the JavaScript logic powers the calculator's functionality. Users can perform addition, subtraction, multiplication, and division, with a user-friendly interface that dynamically updates as they input values and operations. The code is organized and well-documented, making it easy to understand and modify. This calculator website is a versatile and customizable solution that can be easily integrated into other web projects or used as a standalone tool for quick and efficient calculations.